Add helper class for finding interesting offsets.
This helper class is intended to facilitate finding things of interest
nearest to the cursor. The use case is related to context-menu
functionality - if the cursor is in a position which is not interesting
(e.g. whitespace) then search up the AST to find something interesting
(e.g. the class declaration).

The current implementation is quite dumb (but maybe clever enough) - it
traverses the ancestors of the node at the given offset until one is
found which is not an unknown symbol.

Other implementations might query all the refactorings to find out the
closest thing to which one could be applied, or a more strict rule-based
implementation which could return an offset of the first ancestor with
whitelisted node types.